Move data stream resolvability test from IndicesOptionsIntegrationIT to DataStreamIT class. Whether a transport action supports data streams is no longer controlled via indices options.
